create_order_base extracted
