+++ src/core/exchanges/main/mod.rs
@@ -1,5 +1,6 @@
 pub mod cancel_order;
 pub mod create_order;
+pub mod create_order_base;
 pub mod exchange;
 pub mod features;
 pub mod get_open_orders;
